Added Some RE to this Project
I did not use the sources data file but insteaded used my own. In doing so for the last graph as you can see the y axis can not fit all the different formatting for crews. So I learned RE to be able to "CTRL-F" what i needed out of that column leaving that info there and deleting the rest of the column values, then remove the spaces, and finally join back together for better analysis.