Auto search for address on web browser
The runs this code using a batch file. The book shows us how to go into our enviromental setting to
set up the path but it must be out dated. Due to that does not work, and there are about 5-7 different ways
I have seen it done on youtube.

I can make the batch file and click it manually only.