Day 11 - The Blackjack Capstone Project
This was not fun and seems like every area I go to learn there concepts are not gone over as much as needed. 

Ill spend the next 1-2 days going over concepts and viewing this code in https://pythontutor.com/visualize.html#mode=edit until i fully understand it. Its not worth going on if I can not grasp the fundamental basics.